add attribute strategy widget
